Major change in protocol
The protocol has an initialization function `unitnos_procotol_init()`
that sets up a noop signal handler for SIGUSR1.
Communication can be non-blocking (and sometimes must be non-blocking).
In order to avoid busy-loops (which put the CPU at 100%), processes
should call `pause()` and wait for any signal.
`*send*()` functions in unitnos_protocol.h now send SIGUSR1 to the
receiver after the write to pipe with `write()` is finished, so as to
wake up the receiver from its `pause()`.

Previously, orphans processes were unlikely because as soon as the
parent exits, the pipes at the parent side are closed and the next time
the child reads from the pipe, it gets an EOF and terminates in turn.
Now, since every process uses `pause()`, the parent can't just terminate
and expect the child to terminate too, since the child is still waiting
for a signal, due to the call to `pause()`. To remedy this issue,
functions like `unitnos_analyzer_delete`, `unitnos_counter_delete`, etc.
use the function in protocol.h to send a command "close".

By doing so, we're sure that a signal is sent to the child, which wakes
it up from `pause()`.  On reception of this command, the child should
terminate.